requirements.txt
Flask~=3.0.3
pymongo~=4.13.0
python-dotenv~=1.1.0

dnspython~=2.7.0
trio~=0.27.0
idna~=3.10
cryptography~=43.0.3
future~=1.0.0
pip~=25.1.1
attrs~=24.2.0
wheel~=0.45.0
PySocks~=1.7.1
typing_extensions~=4.12.2
MarkupSafe~=3.0.2
Werkzeug~=3.0.4
click~=8.1.7
Jinja2~=3.1.4
blinker~=1.8.2
itsdangerous~=2.2.0
requests~=2.32.3
certifi~=2024.8.30
setuptools~=75.3.0
packaging~=24.2
pyparsing~=3.2.0
pandas~=2.2.3